requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.7
pydantic==2.5.0
asyncio
httpx==0.25.2
